App, Page Component 관심사 분리
